requirements.txt
langgraph>=0.0.15
pypdf>=3.15.1
chromadb>=0.4.18
boto3>=1.28.0
langchain>=0.0.267
langchain-community>=0.0.10
sentence-transformers>=2.2.2
pytesseract>=0.3.10
Pillow>=9.0.0
docx2txt>=0.8
pandas>=1.3.0
openpyxl>=3.0.9
python-pptx>=0.6.21
xlrd>=2.0.1